Fix build errors in daemon code. Also, pgo_get_last_record_number should take a ushort instead of a uint as its record_number parameter.

+++ wh1080/src/main.c
@@ -44,7 +44,7 @@
 BOOL find_range_start(unsigned short *range_start,
                       const unsigned short current_ws_record_id) {
     /* Latest history record available from the database */
-    unsigned int current_db_record_id;
+    unsigned short current_db_record_id;
     time_t current_db_record_timestamp;
     /*history current_db_record_from_ws;*/
